Change parser to use input container instead of template when registering string arguments

+++ demo/main.cpp
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]){
         [](const std::string& str)->int{ return std::stoi(str); },
         [value](const int& i)->bool{ bool valid = i < value ? true : false; return valid; },
         [](const std::exception& e){ std::cout << e.what(); },
-        [&myParser](){ std::cout << "Entered value " << std::to_string(myParser.getValue<int>("positional")) << std::endl; }
+        [&myParser](){ std::cout << "Entered value " << std::to_string((int)myParser.getValue<int>("positional")) << std::endl; }
     );
 
     // Register an option that takes a string and prints it to the console using a callback method.
